refactor: move jwt validation into the shared common package
Consolidate on the portal-assistant variant: async validation off the
event loop, a single long-lived JWKS client, and fail-closed startup —
a missing JWT_JWKS_URL aborts boot unless the dev-only
JWT_INSECURE_ALLOW_UNVERIFIED flag is set (in which case the pod runs
but authenticated routes still reject, matching the previous
request-level behavior). Each agent keeps a thin get_jwt_validator shim
that feeds its own settings into the shared factory.

The observability-plane chart defaults jwksUrl to empty, so it wires
the new flag (default true) to keep default installs booting; setting
a jwksUrl gives full fail-closed validation.
