+++ agents/sre-agent/src/main.py
@@ -11,6 +11,7 @@ from src.api import agent_router, report_router
 from src.auth import check_oauth2_connection, get_oauth2_auth
 from src.auth.dependencies import _load_auth_config
 from src.clients import MCPClient, get_model, get_report_backend
+from src.auth import get_jwt_validator
 from src.config import settings
 from src.logging_config import setup_logging
 from src.mcp_server import drain_background_tasks, make_mcp_app, mcp_server
@@ -26,6 +27,9 @@ if settings.tls_insecure_skip_verify:
 
 @asynccontextmanager
 async def lifespan(_app: FastAPI):
+    logger.info("Initialising JWT validator...")
+    get_jwt_validator()
+
     logger.info("Starting up: Testing LLM connection...")
     try:
         model = get_model()
